demos/serverless-jobs-demo/requirements.txt
# Data Processing Libraries
polars>=0.19.0              # Fast DataFrame library
duckdb>=0.9.0               # In-process SQL OLAP database
pyarrow>=14.0.0             # Apache Arrow for columnar data

# Data Validation & Quality
great-expectations>=0.18.0  # Data validation framework
pandera>=0.17.0             # Statistical data validation

# Healthcare/Medical Data
hl7apy>=1.3.0               # HL7 message parsing
fhir.resources>=7.0.0       # FHIR resource models

# Data Transformation
funcy>=2.0                  # Functional tools for data manipulation
toolz>=0.12.0               # Functional utilities

# Utilities
python-dotenv>=1.0.0        # Environment variable management
tenacity>=8.2.0             # Retry library for robust data pipelines
tqdm>=4.66.0                # Progress bars for long-running operations

# Data Science
scikit-learn>=1.3.0         # Machine learning
statsmodels>=0.14.0         # Statistical models

